requirements.txt
PyYAML==6.0.2
pytest==8.3.5
PySide6==6.8.2.1
prompt_toolkit==3.0.47
